all tests are working: refactoring, fixes and improvements
