+++ settings.template.py
@@ -3,6 +3,7 @@ HUB_URL = 'http://127.0.0.1:4444/wd/hub'
 BROWSER = 'CHROME'  # CHROME or FIREFOX
 
 # initial dialogue with support@liokor.ru must be present
+# all other dialogues and folders of this user would be deleted!
 USERNAME = ''
 PASSWORD = ''
